refactor: rename helm_upgrade command
